api/requirements.txt
Flask>=3.1.1
Flask-SQLAlchemy
Flask-Cors
gunicorn
PyJWT